Refactor: Add JSDoc to Officers component and move officer data to separate file
