+++ src/app/(root)/about/officers/card.tsx
@@ -24,61 +24,47 @@ const OfficerCard: React.FC<CardProps> = ({
   email,
 }) => {
   return (
-    <div className="flex flex-col w-[45%] h-80 items-left justify-center items-center rounded-lg shadow-card padding-10 relative text-center mb-10 shadow-md">
-      <span className="absolute inset-0 bg-gradient-to-r from-[#ec4899] to-[#f472b6] rounded-lg blur-lg opacity-25 group-hover:opacity-40 transition-opacity w-[101%] h-[101%]"></span>
-      <span className="absolute bg-white/50 backdrop-blur-sm border-0 rounded-lg w-[98%] h-[100%]"></span>
+    <div className="relative flex flex-col w-[45%] h-80 rounded-lg text-center mb-10 shadow-card">
+
+      {/* Background gradients */}
+      <span className="absolute inset-0 w-[101%] h-[101%] bg-gradient-to-r from-[#ec4899] to-[#f472b6] rounded-lg blur-lg opacity-25 transition-opacity group-hover:opacity-40"></span>
+      <span className="absolute inset-0 w-[98%] h-full bg-white/50 backdrop-blur-sm rounded-lg"></span>
+
+      {/* Image */}
       <Image
         src={imageUrl}
-        alt={title}
-        className="rounded-[50%] w-32 h-32 shadow-card object-cover z-10"
+        alt={''}
         width={300}
         height={300}
+        className="z-10 w-32 h-32 rounded-full object-cover shadow-card mx-auto mt-4"
       />
-      <div className="card-content z-10">
-        <h2 className="card-title z-10 text-black text-xl font-bold mt-3">
-          {title}
-        </h2>
-        <p className="card-description text-[#f9679d] text-sm font-normal mt-3">
-          {description}
-        </p>
-        <p className="card-description text-[#4B5563] text-sm font-normal mt-3">
-          {major}
-        </p>
-        <div className="flex justify-center mt-3">
-          <a
-            href="mailto: hunterwics@gmail.com"
-            target="_blank"
-            className="px-2"
-          >
-            <Mail
-              href={email}
-              className="hover:scale-150 hover:text-hotpink w-6 h-6  duration-1000 ease-in-out"
-            />
+
+      {/* Content */}
+      <div className="z-10 mt-3">
+        <h2 className="text-xl font-bold text-black">{title}</h2>
+        <p className="mt-2 text-sm font-normal text-[#f9679d]">{description}</p>
+        <p className="mt-1 text-sm font-normal text-[#4B5563]">{major}</p>
+
+        {/* Links */}
+        <div className="flex justify-center mt-3 space-x-4">
+          <a href={`mailto:${email}`} target="_blank" rel="noopener noreferrer">
+            <Mail className="w-6 h-6 hover:scale-150 hover:text-hotpink transition-transform duration-300" />
           </a>
-          <a
-            href="https://www.linkedin.com/company/hunterwics/"
-            target="_blank"
-            className="px-2"
-          >
+          <a href={linkedin} target="_blank" rel="noopener noreferrer">
             <FontAwesomeIcon
               icon={faLinkedin}
-              href={linkedin}
-              className="hover:scale-150 hover:text-hotpink w-6 h-6  duration-1000 ease-in-out"
+              className="w-6 h-6 hover:scale-150 hover:text-hotpink transition-transform duration-300"
             />
           </a>
-          <a
-            href="https://www.linkedin.com/company/hunterwics/"
-            target="_blank"
-            className="px-2"
-          >
+          <a href={github} target="_blank" rel="noopener noreferrer">
             <FontAwesomeIcon
               icon={faGithub}
-              href={github}
-              className="hover:scale-150 hover:text-hotpink w-6 h-6  duration-1000 ease-in-out"
+              className="w-6 h-6 hover:scale-150 hover:text-hotpink transition-transform duration-300"
             />
           </a>
         </div>
       </div>
+
     </div>
   );
 };
